chore: clean up no longer used code
